requirements.txt
Django==5.1.4
dj-database-url==3.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
Pillow==11.3.0
psycopg[binary]==3.2.12
python-dotenv==1.1.1
whitenoise==6.11.0
